Rename meta to SyntaxType
