+++ lib/SyntaxType.js
