+++ peg_src/jsdoctype.pegjs
@@ -1,6 +1,6 @@
 {
   var lodash = require('lodash');
-  var meta = require('../lib/meta.js');
+  var meta = require('../lib/SyntaxType.js');
   var GenericTypeSyntax = meta.GenericTypeSyntax;
   var UnionTypeSyntax = meta.UnionTypeSyntax;
   var NodeType = require('../lib/NodeType.js');
